+++ components/TiptapEditor.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { EditorContent, useCurrentEditor } from '@tiptap/react';
-import { ChangeEventHandler, PropsWithChildren } from 'react';
+import { ChangeEventHandler, PropsWithChildren, useState } from 'react';
 import IconTextAlignLeft from '@/public/icons/IconTextAlignLeft';
 import IconTextAlignMiddle from '@/public/icons/IconTextAlignMiddle';
 import IconTextAlignRight from '@/public/icons/IconTextAlignRight';
@@ -12,8 +12,23 @@ import IconTextItalics from '@/public/icons/IconTextItalics';
 import IconTextNumberPoint from '@/public/icons/IconTextNumberPoint';
 import IconTextUnderline from '@/public/icons/IconTextUnderline';
 import IconTextHighlight from '@/public/icons/IconTextHighlight';
+import { ModalClose, ModalContent, ModalProvider, ModalTrigger } from '@/components/common/Modal';
+import IconAddLink from '@/public/icons/IconAddLink';
+import InputSlid from './common/InputSlid';
+import Button from './common/ButtonSlid';
 
-const TiptapEditor = ({ ...props }: PropsWithChildren) => {
+const TiptapEditor = ({
+  linkUrl,
+  onSaveLinkUrl,
+  ...props
+}: PropsWithChildren<{ linkUrl: string; onSaveLinkUrl: (linkUrl: string) => void }>) => {
+  const [linkUrlValue, setLinkUrlValue] = useState(linkUrl);
+
+  const handleChangeLinkUrlValue: ChangeEventHandler<HTMLInputElement | HTMLSelectElement> = (e) =>
+    setLinkUrlValue(e.target.value);
+
+  const handleSaveLinkUrl = () => onSaveLinkUrl(linkUrlValue);
+
   const { editor } = useCurrentEditor();
 
   if (!editor) return;
@@ -74,9 +89,34 @@ const TiptapEditor = ({ ...props }: PropsWithChildren) => {
             <IconTextColor className='cursor-pointer hover:bg-slate-100' />
           </label>
         </div>
+        <div className='grow flex justify-end'>
+          <ModalProvider>
+            <ModalTrigger type='button'>
+              <IconAddLink className='cursor-pointer hover:bg-slate-100' />
+            </ModalTrigger>
+            <ModalContent className='w-full max-w-[520px] flex flex-col'>
+              <div className='flex justify-between mb-6'>
+                <h1 className='text-lg font-bold'>링크 업로드</h1>
+                <ModalClose />
+              </div>
+              <InputSlid
+                label='링크'
+                type='text'
+                placeholder='영상이나 글, 파일의 링크를 넣어주세요'
+                className='mb-10'
+                value={linkUrlValue}
+                onChange={handleChangeLinkUrlValue}
+              />
+              <ModalClose asChild>
+                <Button className='w-full' onClick={handleSaveLinkUrl}>
+                  확인
+                </Button>
+              </ModalClose>
+            </ModalContent>
+          </ModalProvider>
+        </div>
       </div>
     </>
   );
 };
-
 export default TiptapEditor;
